refactor: split off FissVehicleTeleportationIT from RunFissDrtScenarioIT
